+++ discord-leaderboard/components/Footer.jsx
