+++ discord-leaderboard/components/Heading.jsx
