+++ discord-leaderboard/styles/style.css
@@ -1,17 +1,11 @@
-@import url('https://fonts.googleapis.com/css2?family=Inter&display=swap');
-
-* {
-  font-family: 'Inter', sans-serif;
-  box-sizing: border-box;
-}
-
 /* Nav Area */
 .top-bar {
   width: 100%;
-  margin-top: 0;
+  margin-top: 2%;
   position: relative;
   height: 80px;
   text-align: right;
+  margin-bottom: 2%;
 }
 #title {
   position: absolute;
